requirements.txt
pandas
numpy
scipy
scikit-learn
streamlit
joblib
requests
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv